Add necessary skips to .travis.yml for PyPi
The `skip_cleanup` will fix the removal of the build artifacts
The `skip_existing` will fix the next issue that since you have multiple builds for each python version, they will all try to deploy individually, causing an "existing application" error on PyPi side. This will allow one to succeed and the other to skip if it already exists.